requirements.txt
langchain
langchain-community
groq
pinecone-client
langchain-pinecone
sentence-transformers
streamlit
yfinance
